requirements.txt
fastai==2.7.15
numpy==2.0.0
pandas==2.2.2
Pillow==10.3.0
scikit_surprise==1.1.4
streamlit==1.29.0
surprise==0.1
websocket_client==1.8.0
